python/requirements.txt
# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# JSON-RPC
jsonrpcserver==5.0.9
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
uuid==1.30
